backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.18
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
email-validator>=2.1.1
python-multipart>=0.0.9
